Simulate response data given input lagged env data and an expression

Simulate response data given input lagged env data and an expression

simResponse(x, mu, rfunc = rnorm, ...)

Arguments

x
data.frame containing at least 1 matrix column of lagged data

mu
R expression describing how mean response relates to variables in x

rfunc
function generating response from mean and parameters in ...

...
additional arguments to rfunc

Value

data.frame with all the variables in x plus a response variable

Examples

test <- lagData(env_data, response_data, unit = "month", 6, 24, "envVar")
# generate a response linear in the lag, normal errors sd = 5
test1 <- simResponse(test, 0.05*x1 + rowSums(0.1*lag_6_24*envVar_6_24), sd = 5)
fit1 <- mgcv::gam(response ~ x1 + s(lag_6_24, by = envVar_6_24), data = test1)
summary(fit1)

#> 
#> Family: gaussian 
#> Link function: identity 
#> 
#> Formula:
#> response ~ x1 + s(lag_6_24, by = envVar_6_24)
#> 
#> Parametric coefficients:
#>              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 0.0090493  0.1923729   0.047    0.962    
#> x1          0.0496288  0.0004983  99.592   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Approximate significance of smooth terms:
#>                         edf Ref.df    F p-value    
#> s(lag_6_24):envVar_6_24   2  2.001 2580  <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> R-sq.(adj) =  0.918   Deviance explained = 91.8%
#> GCV = 24.562  Scale est. = 24.499    n = 1564

# poisson response with sine function of lag and an offset
test2 <- simResponse(test, exp(-2 + log(offset) - 0.05*x1 +
rowSums(envVar_6_24*sin(lag_6_24/pi))), rpois)
fit2 <- mgcv::gam(response ~ offset(log(offset))  + x1 + s(lag_6_24, by = envVar_6_24),
data = test2, family = poisson)
summary(fit2)

#> 
#> Family: poisson 
#> Link function: log 
#> 
#> Formula:
#> response ~ offset(log(offset)) + x1 + s(lag_6_24, by = envVar_6_24)
#> 
#> Parametric coefficients:
#>              Estimate Std. Error z value Pr(>|z|)    
#> (Intercept) -2.181114   0.189427  -11.51   <2e-16 ***
#> x1          -0.050297   0.002624  -19.17   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Approximate significance of smooth terms:
#>                           edf Ref.df Chi.sq p-value    
#> s(lag_6_24):envVar_6_24 7.833  8.767   7184  <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> R-sq.(adj) =  0.999   Deviance explained = 99.4%
#> UBRE = -0.94249  Scale est. = 1         n = 1564